feat(pki): Expose `show_certificate_selection_dialog`
Closes #11253
